pyproject.toml
[tool.poetry]
name = "sds-rag"
version = "0.1.0"
description = "A Retrieval-Augmented Generation (RAG) system for analyzing financial documents."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
# Core PDF text and table extraction
pypdf = "^3.17.0"      # For extracting narrative text
tabula-py = "^2.6.0"   # Table extraction (stable alternative)
# Data manipulation and analysis
pandas = "^2.0.0"
numpy = "^1.24.0"
# RAG Framework
langchain = "^0.2.0"
langchain-community = "^0.2.0"
langchain-google-genai = "^1.0.5"
# Vector Database
qdrant-client = "^1.10.0"
# Text embedding
sentence-transformers = "^2.2.2"
# User Interface
streamlit = "^1.32.0"
jpype1 = "^1.6.0"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"